+++ tests/ops/service/privacy_request/test_bigquery_enterprise_privacy_request.py
@@ -91,17 +91,20 @@ def validate_erasure_privacy_request(
     with bigquery_client.connect() as connection:
         stmt = f"select text from enterprise_dsr_testing.post_history where id = {post_history_id};"
         res = connection.execute(stmt).all()
+        assert len(res) == 1
         for row in res:
             assert row.text is None
 
         stmt = f"select user_display_name, text from enterprise_dsr_testing.comments where id = {comment_id};"
         res = connection.execute(stmt).all()
+        assert len(res) == 1
         for row in res:
             assert row.user_display_name is None
             assert row.text is None
 
         stmt = f"select owner_user_id, owner_display_name, body from enterprise_dsr_testing.stackoverflow_posts_partitioned where id = {post_id};"
         res = connection.execute(stmt).all()
+        assert len(res) == 1
         for row in res:
             assert (
                 row.owner_user_id == bigquery_enterprise_resources["user_id"]
@@ -109,18 +112,24 @@ def validate_erasure_privacy_request(
             assert row.owner_display_name is None
             assert row.body is None
 
-        stmt = f"select display_name, location from enterprise_dsr_testing.users where id = {user_id};"
+        stmt = f"select display_name, about_me, location, account_internal from enterprise_dsr_testing.users where id = {user_id};"
         res = connection.execute(stmt).all()
+        assert len(res) == 1
         for row in res:
+            assert row.about_me is None
             assert row.display_name is None
             assert row.location is None
+            # assert nested fields are appropriately handled
+            for item in row.account_internal:
+                assert item["tags"] == []
+                assert item["account_type"] is not None  # not targeted by policy
 
 
 @pytest.mark.integration_bigquery
 @pytest.mark.integration_external
 @pytest.mark.parametrize(
     "dsr_version",
-    ["use_dsr_2_0", "use_dsr_3_0"],
+    ["use_dsr_3_0"],
 )
 @pytest.mark.parametrize(
     "bigquery_fixtures",
@@ -142,7 +151,7 @@ def test_access_request(
     run_privacy_request_task,
     bigquery_enterprise_test_dataset_collections,
 ):
-    request.getfixturevalue(dsr_version)  # REQUIRED to test both DSR 3.0 and 2.0
+    request.getfixturevalue(dsr_version)  # REQUIRED to test DSR 3.0
     request.getfixturevalue(
         bigquery_fixtures
     )  # required to test partitioning and non-partitioned tables
@@ -205,7 +214,7 @@ def test_access_request(
 @pytest.mark.integration_bigquery
 @pytest.mark.parametrize(
     "dsr_version",
-    ["use_dsr_3_0", "use_dsr_2_0"],
+    ["use_dsr_3_0"],
 )
 @pytest.mark.parametrize(
     "bigquery_fixtures",
@@ -225,7 +234,7 @@ def test_erasure_request(
     run_privacy_request_task,
     bigquery_enterprise_test_dataset_collections,
 ):
-    request.getfixturevalue(dsr_version)  # REQUIRED to test both DSR 3.0 and 2.0
+    request.getfixturevalue(dsr_version)  # REQUIRED to test DSR 3.0
     bigquery_enterprise_resources = request.getfixturevalue(bigquery_fixtures)
 
     # first test access request against manually added data